made impassability value represented in persentage
git-svn-id: https://btanks.svn.sourceforge.net/svnroot/btanks/trunk@40 5c0bf991-d514-0410-a65f-ce4832e571f3

+++ tmx/layer.cpp
@@ -1,7 +1,7 @@
 #include "layer.h"
 #include <assert.h>
 
-Layer::Layer(const long w, const long h, const mrt::Chunk & data, const float impassability) : impassability(impassability), _w(w), _h(h){
+Layer::Layer(const long w, const long h, const mrt::Chunk & data, const int impassability) : impassability(impassability), _w(w), _h(h){
 	_data = data;
 }
